tweak console wiring
