+++ src/Domain/Infra/DependencyInjection/ContainerHelper.php
@@ -4,9 +4,7 @@ declare(strict_types=1);
 
 namespace MsgPhp\Domain\Infra\DependencyInjection;
 
-use Doctrine\Bundle\DoctrineBundle\DoctrineBundle;
-use Doctrine\ORM\Version as DoctrineOrmVersion;
-use MsgPhp\Domain\Infra\{Console as ConsoleInfra, SimpleBus as SimpleBusInfra};
+use MsgPhp\Domain\Infra\{SimpleBus as SimpleBusInfra};
 use MsgPhp\Domain\Message\FallbackMessageHandler;
 use SimpleBus\SymfonyBridge\SimpleBusCommandBusBundle;
 use SimpleBus\SymfonyBridge\SimpleBusEventBusBundle;
@@ -75,22 +73,6 @@ final class ContainerHelper
         $definition->setPublic(false);
 
         return $container->setDefinition($class.'.'.ContainerBuilder::hash(__METHOD__.++self::$counter), $definition);
-    }
-
-    public static function registerConsoleClassContextFactory(ContainerBuilder $container, string $class, int $flags = 0): Definition
-    {
-        $definition = self::registerAnonymous($container, ConsoleInfra\Context\ClassContextFactory::class, true)
-            ->setArgument('$class', $class)
-            ->setArgument('$flags', $flags);
-
-        if (class_exists(DoctrineOrmVersion::class) && self::hasBundle($container, DoctrineBundle::class)) {
-            $definition = self::registerAnonymous($container, ConsoleInfra\Context\DoctrineEntityContextFactory::class)
-                ->setAutowired(true)
-                ->setArgument('$factory', $definition)
-                ->setArgument('$class', $class);
-        }
-
-        return $definition;
     }
 
     public static function configureCommandMessages(ContainerBuilder $container, array $classMapping, array $commands): void
